Set custom theme object
In order to allow for custom themes, moves the creation of the
Material-UI theme object into a separate component within the Redux
provider, then applies any supplied configuration settings to the
theme object.
