+++ src/app/src/index.js
@@ -2,31 +2,16 @@
 import 'core-js';
 import React from 'react';
 import { render } from 'react-dom';
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 import { Provider } from 'react-redux';
 import { unregister } from './registerServiceWorker';
 import { store } from './configureStore';
 import './index.css';
 import App from './App';
-import { OARFont } from './util/constants';
-
-const theme = createMuiTheme({
-    typography: {
-        fontFamily: OARFont,
-    },
-    palette: {
-        primary: {
-            main: '#0427a4',
-        },
-    },
-});
 
 /* eslint-disable react/jsx-filename-extension */
 render(
     <Provider store={store}>
-        <MuiThemeProvider theme={theme}>
-            <App />
-        </MuiThemeProvider>
+        <App />
     </Provider>,
     document.getElementById('root'),
 );
